Add files via upload
Attempted to consolidate some of the traces and component placement for the OTTO power board. Only current outstanding item is double checking via sizing for max current load and overall QA check of board. Committing changes to dev branch instead of incorrectly to the master branch (like the first upload).